Move Builder to another file
